refactor(PAT-72): files was moved.
