+++ com/frontend/app/src/containers/CardTicket.jsx
@@ -5,8 +5,8 @@
  **************************************************************************/
 import React, { useEffect } from 'react';
 import { useState } from 'react';
-import BaseTicket from '../components/BaseTicket.jsx';
-import TicketMenu from '../components/TicketMenu.jsx';
+import BaseTicket from '../components/ticket/BaseTicket.jsx';
+import TicketMenu from '../components/ticket/TicketMenu.jsx';
 import { MenuItem } from '@mui/material';
 import DeleteCardDialog from '../components/DeleteCardDialog.jsx';
